requirements-dev.txt
# Development dependencies for churn-risk-scorer
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Type stubs
pandas-stubs>=2.0.0
types-requests>=2.31.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
